+++ src/pretix/control/views/orders.py
@@ -50,7 +50,7 @@ from pretix.base.services.orders import (
 from pretix.base.services.stats import order_overview
 from pretix.base.signals import register_data_exporters
 from pretix.base.templatetags.money import money_filter
-from pretix.base.views.async import AsyncAction
+from pretix.base.views.tasks import AsyncAction
 from pretix.base.views.mixins import OrderQuestionsViewMixin
 from pretix.control.forms.filter import EventOrderFilterForm, RefundFilterForm
 from pretix.control.forms.orders import (
