+++ src/pretix/presale/checkoutflow.py
@@ -26,7 +26,7 @@ from pretix.presale.signals import (
     order_meta_from_request, question_form_fields,
 )
 from pretix.presale.views import CartMixin, get_cart, get_cart_total
-from pretix.presale.views.async import AsyncAction
+from pretix.presale.views.tasks import AsyncAction
 from pretix.presale.views.cart import (
     cart_session, create_empty_cart_id, get_or_create_cart_id,
 )
