--- src/pretix/presale/views/async.py
@@ -1,3 +1,0 @@
-from pretix.base.views.async import *  # noqa
-
-# file kept for backwards compatibility
