+++ src/pretix/presale/views/cart.py
@@ -26,7 +26,7 @@ from pretix.presale.views import (
     EventViewMixin, allow_cors_if_namespaced, allow_frame_if_namespaced,
     iframe_entry_view_wrapper,
 )
-from pretix.presale.views.async import AsyncAction
+from pretix.presale.views.tasks import AsyncAction
 from pretix.presale.views.event import (
     get_grouped_items, item_group_by_category,
 )
