+++ src/pretix/presale/views/tasks.py
@@ -1,0 +1,3 @@
+from pretix.base.views.tasks import *  # noqa
+
+# file kept for backwards compatibility
